+++ react-components/src/Helpers.js
@@ -1,8 +1,12 @@
 /* eslint-disable */
+
 const slugify = function(string) {
   return string.toLowerCase().replace(/ /g,'-').replace(/[^\w-]+/g,'');
 }
 
+
 export {
   slugify,
 }
+
+