+++ react-components/src/components/Modal.jsx
@@ -1,4 +1,3 @@
-/* eslint-disable */
 import React from 'react'
 import PropTypes from 'prop-types'
 import ReactModal from 'react-modal'
@@ -8,7 +7,7 @@ import { withCookies, useCookies } from 'react-cookie';
 
 export function Modal(props){
   const [cookies, setCookie] = useCookies([props.skipFeatureCookieName])
-  
+
   function isOpen() {
     // some modals are opened on user click. Those should be able to skip the
     // "do not open if user previosuly asked not to see the modal again"
@@ -30,9 +29,16 @@ export function Modal(props){
     props.setIsOpen(false)
   }
 
+  function onRequestClose(event) {
+    event.preventDefault()
+    if (!props.preventClose) {
+      props.setIsOpen(false)
+    }
+  }
+
   function getSkipFeature() {
     const SkipFeature = props.skipFeatureComponent
-    if (SkipFeature) {
+    if (SkipFeature && !props.preventClose) {
       return <SkipFeature onClick={handleRequestSkipFeature} />
     }
   }
@@ -40,7 +46,7 @@ export function Modal(props){
   return (
     <ReactModal
       isOpen={isOpen()}
-      onRequestClose={handleClose}
+      onRequestClose={onRequestClose}
       className={'ReactModal__Content ReactModalCentreScreen ' + props.className}
       overlayClassName='ReactModal__Overlay ReactModalCentreScreen'
       contentLabel='Modal'
@@ -59,12 +65,13 @@ Modal.propTypes = {
   skipFeatureCookieName: PropTypes.string,
   id: PropTypes.string,
   performSkipFeatureCookieCheck: PropTypes.bool,
+  preventClose: PropTypes.bool,
 }
 
 Modal.defaultProps = {
   isOpen: false,
   performSkipFeatureCookieCheck: true,
+  preventClose: false,
 }
 
 export default withCookies(Modal)
-