+++ react-components/src/views/ProductsModal/StepCategory.jsx
@@ -9,11 +9,6 @@ import SocialLoginButtons from '@src/components/SocialLoginButtons'
 
 export default function StepCategory(props){
 
-  function handleChange(value) {
-    props.handleChange(value)
-    props.handleSubmit()
-  }
-
   return (
     <div className='great-mvp-export-wizard-step-category'>
       <h2 className="h-m">What are you looking to export?</h2>
@@ -22,7 +17,7 @@ export default function StepCategory(props){
         name="category"
         disabled={props.disabled}
         value={props.value}
-        handleChange={handleChange}
+        handleChange={props.handleChange}
         autofocus={true}
         errors={props.errors.category || []}
         options={[
@@ -39,7 +34,6 @@ StepCategory.propTypes = {
   disabled: PropTypes.bool,
   errors: PropTypes.object,
   handleChange: PropTypes.func.isRequired,
-  handleSubmit: PropTypes.func.isRequired,
   value: PropTypes.string,
 }